// src/__generated__/PaginationListRefetchQuery.graphql.ts
export type PaginationListRefetchQuery$variables = {
  count?: number | null | undefined;
  cursor?: string | null | undefined;
};
export type PaginationListRefetchQuery$data = {
  readonly " $fragmentSpreads": FragmentRefs<"PaginationList_query">;
};
export type PaginationListRefetchQuery = {
  response: PaginationListRefetchQuery$data;
  variables: PaginationListRefetchQuery$variables;
};