// src/__generated__/PaginationList_query.graphql.ts
export type PaginationList_query$data = {
  readonly feed: {
    readonly edges: ReadonlyArray<{
      readonly cursor: string;
      readonly node: {
        readonly author: {
          readonly avatarColor: string;
          readonly name: string;
        };
        readonly id: string;
        readonly likeCount: number;
        readonly title: string;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly totalCount: number;
  };
  readonly " $fragmentType": "PaginationList_query";
};
export type PaginationList_query$key = {
  readonly " $data"?: PaginationList_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"PaginationList_query">;
};